refactor: more things
